requirements.txt
streamlit 
torchvision 
pillow 
matplotlib 
opencv-python-headless 
numpy 
torch 
easyocr 
git+https://github.com/openai/CLIP.git
